+++ hera/main.py
@@ -338,7 +338,7 @@ PICKY_FLAGS = {
     "--stdout": ["assemble"],
 }
 
-VERSION = "hera-py 0.8.0 for HERA version 2.4"
+VERSION = "hera-py 0.8.1 for HERA version 2.4"
 CREDITS = (
     VERSION
     + """